feature: only show active product in selection box

+++ components/forms/SelectBox.tsx
@@ -62,41 +62,45 @@ export const SelectBox = ({
                         leaveTo="opacity-0"
                     >
                         <Listbox.Options className="z-10 absolute mt-1 max-h-60 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-                            {options.map((option, optionId) => (
-                                <Listbox.Option
-                                    key={optionId}
-                                    className={({ active }) =>
-                                        `relative cursor-default select-none py-2 pl-10 pr-4 ${
-                                            active
-                                                ? "bg-amber-100 text-amber-900"
-                                                : "text-gray-900"
-                                        }`
-                                    }
-                                    value={option}
-                                >
-                                    {({ selected }) => (
-                                        <>
-                                            <span
-                                                className={`block truncate ${
-                                                    selected
-                                                        ? "font-medium"
-                                                        : "font-normal"
-                                                }`}
-                                            >
-                                                {option.name}
-                                            </span>
-                                            {selected ? (
-                                                <span className="absolute inset-y-0 left-0 flex items-center pl-3 text-amber-600">
-                                                    <CheckIcon
-                                                        className="h-5 w-5"
-                                                        aria-hidden="true"
-                                                    />
-                                                </span>
-                                            ) : null}
-                                        </>
-                                    )}
-                                </Listbox.Option>
-                            ))}
+                            {options.map((option, optionId) => {
+                                if (option.active) {
+                                    return (
+                                        <Listbox.Option
+                                            key={optionId}
+                                            className={({ active }) =>
+                                                `relative cursor-default select-none py-2 pl-10 pr-4 ${
+                                                    active
+                                                        ? "bg-amber-100 text-amber-900"
+                                                        : "text-gray-900"
+                                                }`
+                                            }
+                                            value={option}
+                                        >
+                                            {({ selected }) => (
+                                                <>
+                                                    <span
+                                                        className={`block truncate ${
+                                                            selected
+                                                                ? "font-medium"
+                                                                : "font-normal"
+                                                        }`}
+                                                    >
+                                                        {option.name}
+                                                    </span>
+                                                    {selected ? (
+                                                        <span className="absolute inset-y-0 left-0 flex items-center pl-3 text-amber-600">
+                                                            <CheckIcon
+                                                                className="h-5 w-5"
+                                                                aria-hidden="true"
+                                                            />
+                                                        </span>
+                                                    ) : null}
+                                                </>
+                                            )}
+                                        </Listbox.Option>
+                                    );
+                                }
+                            })}
                         </Listbox.Options>
                     </Transition>
                 </div>
